The file structure has been split in a more detailed and easier-to-understand way, rather than cramming all function logic into `kernel.c` as done in the os1k tutorial.
At the same time, the process mechanism has been implemented. Currently, the `sscratch` received by all trap handlers is the bottom of the process's kernel stack, so the case where a kernel thread triggers a trap handler is not considered. If an exception is triggered by a kernel thread, it will directly overwrite/clear the contents of the current kernel stack. Similarly, nested exceptions cannot be handled at present.